api/pyproject.toml
[project]
name = "recraft-api"
version = "0.1.0"
description = "Local FastAPI backend for the ReCraft demo"
requires-python = ">=3.12"
dependencies = [
  "boto3>=1.40.0",
  "fastapi>=0.116.1",
  "google-genai>=1.46.0",
  "python-multipart>=0.0.20",
  "python-dotenv>=1.2.2",
  "uvicorn[standard]>=0.35.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
